requirements.txt
streamlit==1.44.1
requests==2.31.0
python-dotenv==1.0.1
pytz==2024.1
openai==1.12.0
httpx==0.26.0 